The 3 Fragments now inflate perfectly in the Minor Activity each time their respective title is clicked.
I also edited the logic for inflating the previous fragments(bottom nav) by remoing the <fragment></fragment> tag in the xml and replacing it with <FrameLayout></FrameLayout> to be consistent with the 3 new fragments way of inflation.

I also setted the URL of Layout Details/Student info in the Main Activity to set the nav drawer texts of student name, major, and college.

Removed setter and getter methods in the Declutterer Class.

Edited the URL method and made it static, and the respective uses of it in the other classes(Fragments and activities).