Removed FrameLayout from all the fragemnts and just kep a relative layout.
Applied a check for login id and pass to desplay errors for user based on specific fault a user did, as well as to determine whether or not to send the json request based on valid parameters or not. (Big help from Anas).

The nav Drawer Header now has the real values of the student (College, Name, and Major). They are requested in the Main Activity, and then setted.